server/requirements.txt
django==5.1.2

djangorestframework==3.15.2
django-filter==24.3
django-colorfield==0.11.0
django-cors-headers==4.4.0
drf-spectacular==0.27.2


# Database
django-environ==0.11.2
psycopg2-binary==2.9.9

# WSGI
gunicorn==23.0.0

# Google Cloud
google-cloud-run==0.10.9
google-auth==2.35.0
django-storages[google]==1.14.4
google-api-python-client==2.149.0

# Testing
pyyaml==6.0.2
model_bakery==1.20.0
Faker==30.3.0
httpx==0.27.2

# Telemetry
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-distro==0.48b0
opentelemetry-exporter-gcp-trace==1.7.0
opentelemetry-instrumentation==0.48b0
opentelemetry-resourcedetector-gcp==1.7.0a0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-instrumentation-django==0.48b0

# end Google code

# user id generation
snowflake-id==1.0.2

# authentication
djangorestframework-simplejwt==5.3.1
